requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.127.1
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyserial==3.5
requests==2.32.5
simpleaudio==1.0.4
sounddevice==0.5.3
srt==3.5.3
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
vosk==0.3.44
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.4
